requirements.txt
pandas
scikit-learn
xgboost
optuna
featurewiz
numpy
scipy
mne
mne-features
librosa
tqdm
pyyaml
pydantic
h5py
vmdpy
PyWavelets
tsaug
joblib